+++ src/Ext/Techno/Body.cpp
@@ -1217,6 +1217,7 @@ void TechnoExt::Serialize(T& Stm)
 		.Process(this->LastTargetCrdClearTimer)
 		.Process(this->ShouldBeDead)
 		.Process(this->PreventCrewEscape)
+		.Process(this->SpawnRandomTarget)
 		;
 }